Move any webdb decorator into main class.
