+++ packages/kartoffelgames.web.database/source/index.ts
@@ -1,6 +1,3 @@
-export { WebDatabaseIdentity } from './web_database/layout/web-database-identity.decorator.ts';
-export { WebDatabaseIndex } from './web_database/layout/web-database-index.decorator.ts';
-export { WebDatabaseMultiIndex } from './web_database/layout/web-database-multi-index.decorator.ts';
 export { WebDatabase } from './web_database/web-database.ts';
 export { WebDatabaseTable } from './web_database/web-database-table.ts';
 export { WebDatabaseTransaction } from './web_database/web-database-transaction.ts';
