+++ packages/kartoffelgames.web.database/source/web_database/query/web-database-query-action.ts
@@ -1,4 +1,4 @@
-import type { TableType } from '../layout/web-database-table-layout.ts';
+import type { TableType } from '../web-database-table-layout.ts';
 import type { WebDatabaseQuery } from './web-database-query.ts';
 
 export class WebDatabaseQueryAction<TTableType extends TableType> {
